Attempt to include tag/value support
While TLV structures are widely used for eg network protocols, they have
a certain overhead due to the use of tags. For some applications, this
overhead is not acceptable. For example, LoRaWAN payloads need to be as
small as possible. As a tradeoff between flexibility and overhead, a TV
structure (tag-value) is often used.

This is an attempt to include TV support in the UTTLV library. It was
recovered from an internal repo that was accidentally removed; all
commit history is lost and there's a lot of unneeded whitespace changes.

This path was abandoned and a different TV decoder was written for
internal use.